% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.Boucher2016.R
\docType{data}
\name{dat.Boucher2016}
\alias{dat.Boucher2016}
\title{PARESTHESIA RATES WITH TOPIRAMATE IN MIGRAINE PROPHYLAXIS TRIALS}
\format{
A data frame with following coloumns
\describe{
  \item{d1}{dose (mg) in the first arm (placebo)}
  \item{r1}{number of events in the first arm (placebo)}
  \item{n1}{number of patients in the first arm (placebo)}
  \item{d2}{dose (mg) in the second arm }
  \item{r2}{number of events in the second arm}
  \item{n2}{number of patients in the second arm}
  \item{d3}{dose (mg) in the third arm}
  \item{r3}{number of events in the third arm}
  \item{n3}{number of patients in the third arm}
  \item{d4}{dose (mg) in the fourth arm}
  \item{r4}{number of events in the fourth arm}
  \item{n4}{number of patients in the fourth arm}
  \item{nd}{number of treatment arms}
 }
}
\source{
Boucher M, and Bennets M. The Many Flavors of
Model-Based Meta-Analysis: Part I - Introduction and Landmark Data.
CPT Pharmacometrics Syst. Pharmacol. (2016) 5, 54-64; doi:10.1002/psp4.12041
}
\usage{
dat.Boucher2016
}
\description{
Numbers of patients and events (paresthesia rates)
in experimental and control groups of 6 studies.
It is in one-study-per-row format.
}
\keyword{datasets}
