% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot.R
\name{forestplot.meta_stan}
\alias{forestplot.meta_stan}
\title{Plot a forest plot}
\source{
This function is based \code{foresplot} function from \code{foresplot}
R package.
}
\usage{
\method{forestplot}{meta_stan}(x = NULL, data_wide = NULL, digits = 2, boxsize = 0.3, col, ...)
}
\arguments{
\item{x}{A \code{meta_stan} object.}

\item{data_wide}{Dataset in a wide format, one-row-per-study.}

\item{digits}{A numerical value specifying the number of significant digits to be shown.
Default is 2.}

\item{boxsize}{A numerical value specifying the box size. Default is 0.3.}

\item{col}{A function specifying the colors. See \code{forestplot::fpColors} for details.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The return value is invisible \code{NULL}.
#' @examples
\dontrun{
data('dat.Crins2014', package = "MetaStan")
dat_long <- convert_data_arm(dat.Crins2014$exp.total, dat.Crins2014$cont.total,
                             dat.Crins2014$exp.AR.events, dat.Crins2014$cont.AR.events)
bnhm.Crins  <- meta_stan(data = dat_long, family = "binomial",
                         mu_prior = c(0, 10), theta_prior = c(0, 100),
                         tau_prior =  0.5)
forestplot(bnhm.Crins)

}
}
\description{
Takes a \code{meta_stan} object which is obtained by function \code{meta_stan} and plot
a forestplot, showing individual estimates along with their 95 percent credible intervals,
resulting effect estimate and prediction interval.
}
\seealso{
\code{foresplot::foresplot}
}
\author{
Christian Roever and Burak Kuersad Guenhan
}
