% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.MBMA_stan.R
\name{print.MBMA_stan}
\alias{print.MBMA_stan}
\title{Print MBMA object}
\usage{
\method{print}{MBMA_stan}(x, digits = 2, ...)
}
\arguments{
\item{x}{A \code{MBMA_stan} object.}

\item{digits}{An integer indicating the number of decimal places.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The return value is invisible \code{NULL}
}
\description{
Takes an \code{MBMA_stan} object which is obtained by function \code{MBMA_stan} and print
the model and data information such as model type used in the model.
}
