% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.meta_stan}
\alias{print.meta_stan}
\title{Print meta_stan object}
\usage{
\method{print}{meta_stan}(x, digits = 2, ...)
}
\arguments{
\item{x}{A \code{meta_stan} object.}

\item{digits}{An integer indicating the number of decimal places.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The return value is invisible \code{NULL}
}
\description{
Takes an \code{meta_stan} object which is obtained by function \code{meta_stan} and print
the model and data information such as model type used in the model.
}
\details{
The resulting data.frame can be used as data argument in \code{meta_stan}.
}
