% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_arm.R
\name{convert_data_arm}
\alias{convert_data_arm}
\title{Convert contrast-based dataset to arm-based dataset}
\usage{
convert_data_arm(nt, nc, pt, pc, pub, data = NULL)
}
\arguments{
\item{nt}{Number of subjects in treatment arm}

\item{nc}{Number of subjects in control arm}

\item{pt}{Number of events in treatment arm}

\item{pc}{Number of events in treatment arm}

\item{pub}{The corresponding publication}

\item{data}{Optional data frame containing the variables given to the arguments above.}
}
\value{
A dataframe object
}
\description{
\code{convert_data_arm} creates a dataframe to fit a meta-analysis
model using \code{meta_Stan} function.
}
\examples{
## Create arm-based dataset
data('dat.Crins2014', package = "MetaStan")
dat_converted <- convert_data_arm(exp.total, cont.total,
                                  exp.AR.events, cont.AR.events,
                                  publication, data = dat.Crins2014)


}
