% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_stan.R
\name{meta_stan}
\alias{meta_stan}
\title{Bayesian meta-analysis with Stan}
\usage{
meta_stan(stanDat)
}
\arguments{
\item{stanDat}{Data}
}
\value{
an object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian meta-analysis with Stan
}
