% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_stan.R
\name{meta_stan}
\alias{meta_stan}
\title{Fitting a meta-analysis model using Stan}
\usage{
meta_stan(ntrt = NULL, nctrl = NULL, rtrt = NULL, rctrl = NULL,
  mu_prior = c(0, 10), theta_prior = NULL, tau_prior = c(0, 0.5),
  tau_prior_dist = "half-normal", delta_u = NULL)
}
\arguments{
\item{ntrt}{Number of subjects in treatment arm}

\item{nctrl}{Number of subjects in control arm}

\item{rtrt}{Number of events in treatment arm}

\item{rctrl}{Number of events in contrl arm}

\item{mu_prior}{A numerical vector specifying the parameter of the normal prior
density for baseline risks, first value is parameter for mean, second is for variance.
Default is c(0, 10).}

\item{theta_prior}{A numerical vector specifying the parameter of the normal prior
density for treatment effect estimate, first value is parameter for mean, second
is for variance. Default is NULL.}

\item{tau_prior}{A numerical vector specifying the parameter of the prior density
for heterogenety stdev. Defalut is c(0, 0.5).}

\item{tau_prior_dist}{A string specifying the prior density for the heterogeneity standard deviation,
option is 'half-normal' for half-normal prior.}

\item{delta_u}{A numerical value specifying the upper bound of the a priori inerval for
treatment effect on odds ratio scale. This is used to cacluate a normal weakly informative prior
for theta. Thus when this argument is pecified, `theta` should be left empty. Default is NULL.}
}
\value{
an object of class `stanfit` returned by `rstan::sampling`
}
\description{
`meta_stan` fits a meta-analysis model using Stan.
}
\examples{
\dontrun{
data('dat.Crins2014', package = "MetaStan")
## Subset of dataset where PTLD outcomes available
dat.Crins2014.PTLD = subset(dat.Crins2014, is.na(exp.PTLD.events) == FALSE)
## Fitting a Binomial-Normal Hierarchial model
bnhm.vague.PTLD.stan  <- meta_stan(ntrt = dat.Crins2014.PTLD$exp.total,
                                   nctrl = dat.Crins2014.PTLD$cont.total,
                                   rtrt = dat.Crins2014.PTLD$exp.PTLD.events,
                                   rctrl = dat.Crins2014.PTLD$cont.PTLD.event,¨
                                   mu_prior = c(0, 10), theta_prior = c(0, 100),
                                   tau_prior = c(0, 0.5))
}

}
