% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MBMA_stan.R
\name{plot.MBMA_stan}
\alias{plot.MBMA_stan}
\title{Plot a dose-response plot}
\source{
This function uses \code{ggplot} function from \code{ggplot2}
R package.
}
\usage{
\method{plot}{MBMA_stan}(x = MBMA.stan, ...)
}
\arguments{
\item{x}{A \code{MBMA_stan} object.}

\item{...}{Further arguments passed to ggplot.}
}
\value{
The return value is invisible \code{NULL}.
}
\description{
Takes a \code{MBMA_stan} object which is obtained by function \code{MBMA_stan} and plot
a dose-response plot, showing observed event probabilities and the estimated dose-response
function with pointwise 95% credible intervals.
}
\examples{
\dontrun{
## Load the dataset
data('dat.Eletriptan', package = "MetaStan")
## Fitting a Binomial-Normal Hierarchial model using WIP priors
datMBMA = create_MBMA_dat(dat = dat.Eletriptan,
                         armVars = c(dose = "d", r = "r",
                                     n = "n"),
                         nArmsVar = "nd")

MBMA.Emax  <- MBMA_stan(data = datMBMA,
                          family = "binomial",
                          dose_response = "emax",
                          Emax_prior = c(0, 10),
                          ED50_prior = "functional",
                          tau_prior_dist = "half-normal",
                          tau_prior = 0.5)
print(MBMA.Emax)
plot(MBMA.Emax)
# use ggplot2 package for further adjustments
Library(ggplot2)
theme_set(theme_classic())
plot(MBMA.Emax) + xlab("Doses (mg)") + ylab("response probabilities")

}
}
\seealso{
\code{ggplot2::ggplot}
}
\author{
Christian Roever and Burak Kuersad Guenhan
}
